# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Fitness RPG API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /equipment:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: List equipment
      responses:
        "200":
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: addEquipment
      tags: [equipment]
      summary: Add equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Added equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Remove equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SKILLS ──────────────────────────────────────────────────────────────────
...
  /inventory/{id}/use:
    post:
      operationId: useInventoryItem
      tags: [inventory]
      summary: Use an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Use result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UseItemResult"

  # ─── STORE ────────────────────────────────────────────────────────────────────
...
  /store/purchase:
    post:
      operationId: purchaseStoreItem
      tags: [store]
      summary: Purchase a store item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "200":
          description: Purchase result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResult"

  # ─── ANALYTICS ────────────────────────────────────────────────────────────────
...
  /analytics/volume:
    get:
      operationId: getVolumeByMuscleGroup
      tags: [analytics]
      summary: Training volume by muscle group
      parameters:
        - name: weeks
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Volume data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MuscleGroupVolume"
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, name, level, rank, xp, xpToNextLevel, hp, maxHp, mp, maxMp, gold, freeStatPoints, stats, activeTitle, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        rank:
          type: string
          enum: [E, D, C, B, A, S, "National-Level"]
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        mp:
          type: integer
        maxMp:
          type: integer
        gold:
          type: integer
        freeStatPoints:
          type: integer
        streakDays:
          type: integer
        stats:
          $ref: "#/components/schemas/PlayerStats"
        activeTitle:
          type: ["string", "null"]
        penaltyQuestActive:
          type: boolean
        createdAt:
          type: string
    PlayerStats:
      type: object
      required: [strength, agility, stamina, vitality, discipline, sense]
      properties:
        strength:
          type: integer
        agility:
          type: integer
        stamina:
          type: integer
        vitality:
          type: integer
        discipline:
          type: integer
        sense:
          type: integer
...
    Equipment:
      type: object
      required: [id, name, category, owned, available]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [barbell, rack, machine, free_weights, cardio, mat, striking, cable, bench, other]
        description:
          type: ["string", "null"]
        owned:
          type: boolean
        available:
          type: boolean
        maxWeight:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    EquipmentInput:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
        category:
          type: string
          enum: [barbell, rack, machine, free_weights, cardio, mat, striking, cable, bench, other]
        description:
          type: ["string", "null"]
        owned:
          type: boolean
        available:
          type: boolean
        maxWeight:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        owned:
          type: boolean
        available:
          type: boolean
        maxWeight:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    # ─── SKILLS ────────────────────────────────────────────────────────────────
...
    UseItemResult:
      type: object
      required: [success, message, player]
      properties:
        success:
          type: boolean
        message:
          type: string
        player:
          $ref: "#/components/schemas/Player"
        bonusReward:
          type: ["string", "null"]

    # ─── STORE ─────────────────────────────────────────────────────────────────
...
    PurchaseInput:
      type: object
      required: [itemId]
      properties:
        itemId:
          type: integer
        quantity:
          type: integer
    PurchaseResult:
      type: object
      required: [success, message, goldSpent, remainingGold, player]
      properties:
        success:
          type: boolean
        message:
          type: string
        goldSpent:
          type: integer
        remainingGold:
          type: integer
        player:
          $ref: "#/components/schemas/Player"

    # ─── ANALYTICS ─────────────────────────────────────────────────────────────
...
    MuscleGroupVolume:
      type: object
      required: [muscleGroup, totalSets, totalReps, totalVolume]
      properties:
        muscleGroup:
          type: string
        totalSets:
          type: integer
        totalReps:
          type: integer
        totalVolume:
          type: number
...